Separate array and object extensions from id3_tree
