+++ lib/decisiontree.rb
@@ -1,2 +1,3 @@
 require File.dirname(__FILE__) + '/decisiontree/id3_tree.rb'
-require 'core_extension/array.rb'
+require 'core_extensions/object'
+require 'core_extensions/array'
